За старт взял проект с предыдущего урока. Переименовал в Lesson_4 для сохранения ясности.
