+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "Lesson_3"
+rootProject.name = "Lesson_4"
